Refactor book update for author changes and change some tests
All tests are now green. Really important that the tests are built in a way where there is minimal dependencies between tests

+++ Controllers/bookController.js
@@ -5,7 +5,7 @@ const format = require('pg-format');
 const path = require('path');
 
 const bookCoverPath = path.join(__dirname, '../public/uploads/bookCovers/');
-const expectedRequestKeys = ['title', 'pages', 'date_published', 'cover', 'author_ids'];
+const expectedRequestKeys = [...Object.keys(Book.rawAttributes), 'author_ids'];
 
 // Creates a new book entry in the database
 const postBook = async (req, res) => {
@@ -89,15 +89,27 @@ const linkBookToAuthors = async (book, authorIds) => {
 };
 
 // Remove one or more authors from the given book
-const removeAuthorsFromBook = async (bookID, authorsToRemove) => {
+const removeAuthorsFromBook = async (book, authorsToRemove) => {
   if (authorsToRemove.length === 0) {
     return {};
   }
 
-  const query = format(`DELETE FROM book_author WHERE book_id = ${bookID} AND author_id IN (%L) RETURNING *`, authorsToRemove);
-  const removed = await pool.query(query, []);
+  const successfulRemovals = [];
+  for(const authorId of authorsToRemove){
+    const authorToRemove = await Author.findOne({where: {id: authorId}});
 
-  return removed;
+    if(isNullOrEmpty(authorToRemove)){
+      continue;
+    }
+
+    await book.removeAuthor(authorToRemove);
+    successfulRemovals.push(authorId);
+  }
+  
+  // const query = format(`DELETE FROM book_author WHERE book_id = ${book} AND author_id IN (%L) RETURNING *`, authorsToRemove);
+  // const removed = await pool.query(query, []);
+
+  return successfulRemovals;
 };
 
 // Update book entry with the values in the request body
@@ -106,6 +118,11 @@ const updateBook = async (req, res) => {
   try {
     if (isNullOrEmpty(req.body)) {
       throw new Error('The body is empty');
+    }
+
+    let book = await findBook({id: bookId});
+    if (isNullOrEmpty(book.dataValues)) {
+      throw new Error(`Book with id = ${bookId} not found`);
     }
 
     // Use the one below if cover image uploads is included with the other details
@@ -122,14 +139,8 @@ const updateBook = async (req, res) => {
 
     // Check if there's a new image sent in the request
     if (req.file != null) {
-      // Get entry to update
-      const book = await findBook({id: bookId});
-      if (isNullOrEmpty(book)) {
-        throw new Error(`Book with id = ${bookId} not found`);
-      }
-
       // Delete old picture
-      deleteFile(`${bookCoverPath}${book.cover}`);
+      if(book.cover != null) deleteFile(`${bookCoverPath}${book.cover}`);
 
       // Gotta change the book's cover image with the new filepath
       bookChanges.cover = req.file.filename;
@@ -138,21 +149,20 @@ const updateBook = async (req, res) => {
     // Update the general details of the book
     let successfulUpdate = false;
 
-    if (bookChanges != null) {
-      await changeBookDetails(bookChanges, bookId);
+    if (!isNullOrEmpty(bookChanges)) {
+      book = await changeBookDetails(bookChanges, book);
       successfulUpdate = true;
     }
 
-    if (authorChange != null) {
+    if (!isNullOrEmpty(authorChange)) {
       // Update the author if the body specified it
-      await changeAuthor(bookId, authorChange.authorsToRemove, authorChange.authorsToAdd);
-      // console.log("Author changed");
+      book = await changeAuthor(book, authorChange.authorsToRemove, authorChange.authorsToAdd);
       successfulUpdate = true;
     }
 
     if (successfulUpdate) {
       // Can't return this if its just the author that is changed
-      res.status(200).json('Successful update');
+      res.status(200).json(book);
     } else {
       res.status(400).json('The request body contains the wrong keys');
     }
@@ -167,61 +177,58 @@ const updateBook = async (req, res) => {
 };
 
 // Updates the specified book's general details (basically all details except the author)
-const changeBookDetails = async (updateBody, bookId) => {
-  if (Object.keys(updateBody).length === 0) {
-    throw new Error('No details to change are specified');
+const changeBookDetails = async (updateBody, book) => {
+  if(!validateRequestBody(Book, updateBody)){
+    throw new Error(`The keys for model: ${Book.name} are wrong`);
   }
-
-  const [ rowsUpdated, [updatedBook] ] = await Book.validatedUpdate(updateBody, {returning: true, where: {id: bookId}});
-  console.log(updatedBook);
+  
+  const updatedBook = await book.update(updateBody);
 
   return updatedBook.dataValues;
 };
 
 // A function that allows changing the author(s) of a book
-const changeAuthor = async (bookID, authorsToRemove = [], authorsToAdd = []) => {
-  // Get the book in question
-  const bookToUpdate = await pool.query('SELECT * FROM books WHERE id = ($1)', [bookID]);
-  if (isNullOrEmpty(bookToUpdate)) {
-    throw new Error(`Book with id = ${bookID} not found`);
+const changeAuthor = async (bookToUpdate, authorsToRemove = [], authorsToAdd = []) => {
+  if (isNullOrEmpty(bookToUpdate.dataValues)) {
+    throw new Error(`Given book to update is null`);
   };
 
   // Get the author(s) of the book
-  const bookAuthors = await getBookAuthor(bookID);
-  const bookAuthorIDs = bookAuthors.map((author) => author.id);
+  // const bookAuthors = await getBookAuthor(bookID);
+  // const bookAuthorIDs = bookAuthors.map((author) => author.id);
+
+  const bookAuthorIds = bookToUpdate.Authors.map((author) => author.dataValues.id);
+  console.log(`authorsToRemove: ${authorsToRemove}\nbookAuthorIds: ${bookAuthorIds}`);
 
   // Checks whether all ids in authorsToRemove is present in the array of authors for this book
-  if (authorsToRemove.length > 0 && !checkArrayContent(bookAuthorIDs, authorsToRemove)) {
+  if (authorsToRemove.length > 0 && !checkArrayContent(bookAuthorIds, authorsToRemove)) {
     throw new Error('Invalid author ids to remove');
   }
+
+  console.log(bookAuthorIds);
 
   // Verify whether the author ids to add are authors that exist in the database
   if (authorsToAdd.length > 0) {
     // Look for the author in the authors table using their id
     const authorValid = await checkAuthorPresence(authorsToAdd);
-    console.log(authorValid);
-    // Make sure that none of the authors are already credited with writing this book
-
+    
     if (!authorValid) {
       throw new Error('At least one of the authors to add don\'t exist in the database');
     }
 
     // If at least one of the given authors to add has been credited with writing this book, then you can't add these authors
-    if (!checkUniqueness(bookAuthorIDs, authorsToAdd)) {
+    if (!checkUniqueness(bookAuthorIds, authorsToAdd)) {
       throw new Error('At least one of the authors is already credited with writing this book');
     }
   }
 
   // Delete corresponding book_author entry
-  const removed = await removeAuthorsFromBook(bookID, authorsToRemove);
+  const removed = await removeAuthorsFromBook(bookToUpdate, authorsToRemove);
 
   // Create new entries in the book_author table
-  const added = await linkBookToAuthors(bookID, authorsToAdd);
+  const added = await linkBookToAuthors(bookToUpdate, authorsToAdd);
 
-  return {
-    removed: removed.rows,
-    added: added.rows,
-  };
+  return added;
 };
 
 
@@ -385,6 +392,7 @@ const findBook = async (queryConditions) => {
 const validateBookRequestBody = (body) => {
   let isValid = true;
   let areAuthorsSpecified = false;
+  
   Object.keys(body).forEach(key => {
     // Find whether key exists in specified key list
     if(!expectedRequestKeys.includes(key)) isValid = false;
